Refactored Order service to no longer inject DbContext
