+++ Turnup/Services/OrderService/OrderService.cs
@@ -38,8 +38,59 @@ public class OrderService : IOrderService
 
     }
 
-    public async Task<ServiceResponse<Cart>> PlaceOrder(Claim? user)
+    public async Task<ServiceResponse<Order>> PlaceOrder(Claim? user)
     {
-      return await _cartService.GetCart(user);
+      var cart= await _cartService.GetCart(user);
+      var order = Order(cart);
+
+      return new ServiceResponse<Order>()
+      {
+          Data = order.Result
+      };
+    }
+    
+    private async Task<Order> Order(ServiceResponse<Cart> cart)
+    {
+        var order = new Order
+        {
+            CustomerId = cart.Data.CustomerId,
+            EstablishmentId = cart.Data.EstablishmentId,
+            OrderDate = DateTime.Now,
+            OrderItems = new List<OrderItem>(),
+            Status = OrderStatus.Pending.ToString(),
+            SubTotal = cart.Data.Subtotal,
+        };
+
+        order.OrderItems.AddRange(MapCartItemToOrderItem(cart.Data.Items));
+        order.Total = order.CalculateTotal();
+
+        await _context.Orders.AddAsync(order);
+        _context.Carts.Remove(cart.Data);
+        await _context.SaveChangesAsync();
+        return order;
+    }
+    
+    private List<OrderItem> MapCartItemToOrderItem(List<CartItem> cartItems)
+    {
+        List<OrderItem> items = new List<OrderItem>();
+        
+        foreach (var item in cartItems)
+        {
+            var orderItem = new OrderItem
+            {
+                EstablishmentId = item.Product.EstablishmentId,
+                CustomerId = item.Cart.CustomerId,
+                Product = item.Product,
+                Title = item.Product.Title,
+                Quantity = item.Quantity,
+                Price = item.Product.Price,
+                ProductId = item.ProductId
+
+            };
+
+            items.Add(orderItem);
+        }
+
+        return items;
     }
 }
